Feature: self-upgrade: for unification tokio::spwn file renamed to server.rs
